EE-8231 Add check box for asking question as proxy
The checkbox this commit adds is a small unstyled checkbox and not the
large GDS checkbox used on other Government services.

This is due to a bug in HOF's template mixins: for more information see
https://github.com/UKHomeOfficeForms/hof-template-mixins/pull/27

We're supposed to be able to override the problematic partial by
overriding it in apps/common/views/partials/forms/checkbox.html but it
doesn't seem to work.