Fix proposition header link
The proposition header looks for `baseUrl`, but we no longer set that
because our app now lives at the root.

Mustache and W3C standards can’t cope with that, so defaults to relative
root (i.e. the same page). This commit overrides the navigation partial
and hardcodes the href in the proposition header to /.